Rename : `ArticleFeed` to `FullArticleView`, `MyArticles` to `MyArticlesView`
