--- frontend/src/pages/ArticleFeedPage.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ArticleFeedContainer from '../containers/article/ArticleFeedContainer';
-
-const ArticleFeedPage = () => {
-	return <ArticleFeedContainer />;
-};
-
-export default ArticleFeedPage;
