+++ frontend/src/pages/FullArticlesPage.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import FullArticlesContainer from '../containers/article/FullArticlesContainer';
+
+const FullArticlesPage = () => {
+	return <FullArticlesContainer />;
+};
+
+export default FullArticlesPage;
